集成登录拦截器，使用Jwt进行拦截。

+++ record-account-service/src/main/java/cn/rollin/bean/model/ClassifyDO.java
@@ -1,16 +1,17 @@
 package cn.rollin.bean.model;
 
-import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.annotation.IdType;
-import java.util.Date;
 import com.baomidou.mybatisplus.annotation.TableId;
-import java.io.Serializable;
+import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+import java.io.Serializable;
+import java.util.Date;
+
 /**
  * <p>
- * 
+ *
  * </p>
  *
  * @author rollin
@@ -32,7 +33,7 @@ public class ClassifyDO implements Serializable {
     /**
      * 用户ID
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 收入分类ID集合  1,2,3,4,5,6...
